Implemented castling, startpos perft accurate up until depth 7 (didn't test 8 yet). Seem to be either ep or castling issues on other positions that need fixing.

+++ src/move.h
@@ -81,7 +81,7 @@ inline Move encodeMove(const uint8_t fromSquare, const uint8_t toSquare, const P
 }
 
 inline uint8_t getFromSquare(const Move move) {
-    return static_cast<uint8_t>(move & 0x3F);
+    return move & 0x3F;
 }
 
 inline uint8_t getToSquare(const Move move) {
